pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insider-trading-detector"
version = "0.1.0"
description = "Detect potential insider trading on prediction markets (Polymarket, Kalshi)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Prediction Markets Research"}
]
keywords = ["prediction-markets", "polymarket", "kalshi", "insider-trading", "analytics"]

dependencies = [
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "web3>=6.0.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
insider-detector = "insider_detector.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
